requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.0.7
django-cuser==2017.3.16
django-debug-toolbar==4.4.6
django-filter==24.3
django-modeltranslation==0.19.9
django-redis==5.4.0
django-templated-mail==1.1.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docutils==0.21.2
drf-spectacular==0.27.2
flake8==7.1.1
flake8-docstrings==1.7.0
google-trans-new==1.1.9
gunicorn==23.0.0
hiredis==3.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.1
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.9.0
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
